+++ Ws2812LedController.AudioReactive/Effects/Volume/LightUpReactiveEffect.cs
@@ -16,7 +16,7 @@ public class LightUpReactiveEffect : BaseAudioReactiveEffect, IHasVolumeAnalysis
     public override string FriendlyName => "Light up";
     public override string Description => "Light single LEDs based on volume peaks up";
     public override int Speed { set; get; } = 1000 / 60;
-    public int FadeSpeed { set; get; } = 6;
+    public byte FadeSpeed { set; get; } = 6;
     public IVolumeAnalysisOption VolumeAnalysisOptions { set; get; } = new AgcVolumeAnalysisOption(128);
     public double Threshold { set; get; } = 1500;
